replaced int with size_t

+++ BinarySearch.cpp
@@ -1,9 +1,9 @@
 #include "BinarySearch.h"
 
 // Recursive Binary Search Implementation
-int recursiveBinarySearch(const std::vector<int>& vec, int target, int left, int right) {
+int recursiveBinarySearch(const std::vector<int>& vec, int target, size_t left, size_t right) {
     if (right >= left) {
-        int mid = left + (right - left) / 2;
+        size_t mid = left + (right - left) / 2;
         if (vec[mid] == target) return mid;
         if (vec[mid] > target) return recursiveBinarySearch(vec, target, left, mid - 1);
         return recursiveBinarySearch(vec, target, mid + 1, right);
@@ -13,9 +13,9 @@ int recursiveBinarySearch(const std::vector<int>& vec, int target, int left, int
 
 // Iterative Binary Search Implementation
 int iterativeBinarySearch(const std::vector<int>& vec, int target) {
-    int left = 0, right = vec.size() - 1;
+    size_t left = 0, right = vec.size() - 1;
     while (left <= right) {
-        int mid = left + (right - left) / 2;
+        size_t mid = left + (right - left) / 2;
         if (vec[mid] == target) return mid;
         if (vec[mid] < target) left = mid + 1;
         else right = mid - 1;
