+++ BinarySearch.h
@@ -4,7 +4,7 @@
 #include <vector>
 
 // Function prototypes
-int recursiveBinarySearch(const std::vector<int>& vec, int target, int left, int right);
+int recursiveBinarySearch(const std::vector<int>& vec, int target, size_t left, size_t right);
 int iterativeBinarySearch(const std::vector<int>& vec, int target);
 int sequentialSearch(const std::vector<int>& vec, int target);
